+++ Source/Krypton Components/Krypton.Workspace/Workspace/KryptonWorkspaceSequenceDesigner.cs
@@ -16,7 +16,9 @@ internal class KryptonWorkspaceSequenceDesigner : ComponentDesigner
 {
     #region Instance Fields
     private KryptonWorkspaceSequence? _sequence;
+    private IDesignerHost? _designerHost;
     private IComponentChangeService? _changeService;
+    private DesignerVerbCollection? _verbs;
     #endregion
 
     #region Identity
@@ -42,8 +44,7 @@ internal class KryptonWorkspaceSequenceDesigner : ComponentDesigner
 
         // Cast to correct type
         _sequence = component as KryptonWorkspaceSequence;
-
-        // Get access to the services
+        _designerHost = GetService(typeof(IDesignerHost)) as IDesignerHost;
         _changeService = GetService(typeof(IComponentChangeService)) as IComponentChangeService;
 
         // We need to know when we are being removed/changed
@@ -64,6 +65,26 @@ internal class KryptonWorkspaceSequenceDesigner : ComponentDesigner
             compound.AddRange(_sequence?.Children!);
 
             return compound;
+        }
+    }
+
+    /// <summary>
+    /// Gets the design-time verbs shown on the component context menu.
+    /// </summary>
+    public override DesignerVerbCollection Verbs
+    {
+        get
+        {
+            if (_verbs == null)
+            {
+                _verbs = new DesignerVerbCollection
+                {
+                    new DesignerVerb(@"Add Cell", OnAddCell),
+                    new DesignerVerb(@"Add Sequence", OnAddSequence)
+                };
+            }
+
+            return _verbs;
         }
     }
     #endregion
@@ -137,5 +158,11 @@ internal class KryptonWorkspaceSequenceDesigner : ComponentDesigner
             }
         }
     }
+
+    private void OnAddCell(object? sender, EventArgs e) =>
+        KryptonWorkspaceDesignerActions.AddCell(_sequence?.WorkspaceControl, _sequence?.Children, _designerHost, _changeService);
+
+    private void OnAddSequence(object? sender, EventArgs e) =>
+        KryptonWorkspaceDesignerActions.AddSequence(_sequence?.WorkspaceControl, _sequence, _sequence?.Children, _designerHost, _changeService);
     #endregion
 }